Extract identity and maintenance services
